# label2rgb performance fix